requirements.txt

flask
transformers
torch
easyocr
sentence-transformers
langdetect
pillow
opencv-python-headless
